git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@53716 954022d7-b5bf-4e40-9824-e11837661b57

+++ DesktopEditor/graphics/ImageFilesCache.h
@@ -94,7 +94,7 @@ public:
 			return pImage;
 		}
 
-		int nCount = m_mapImages.size();
+		int nCount = (int)m_mapImages.size();
 		if (nCount >= m_lMaxCount)
 		{
 			int nNeedDelete = nCount - m_lMaxCount;
